requirements.txt
pyyaml==6.0.2
dotenv==0.9.9
pytest==8.4.1
#pytest-mock
